+++ packages/moonstone/internal/Picker/PickerButton.js
@@ -1,4 +1,4 @@
-import Holdable from '@enact/ui/Holdable';
+import Holdable from '../Holdable';
 import kind from '@enact/core/kind';
 import React from 'react';
 import PropTypes from 'prop-types';