+++ packages/sampler/stories/qa-stories/Input.js
@@ -20,6 +20,7 @@ storiesOf('Input')
 		'with long text',
 		() => (
 			<Input
+				autoFocus={boolean('autoFocus')}
 				onChange={action('onChange')}
 				disabled={boolean('disabled')}
 				iconAfter={select('iconAfter', iconNames)}
@@ -38,6 +39,7 @@ storiesOf('Input')
 		() => (
 			<div>
 				<Input
+					autoFocus={boolean('autoFocus')}
 					onChange={action('onChange')}
 					disabled={boolean('disabled')}
 					iconAfter={select('iconAfter', iconNames)}
@@ -48,6 +50,7 @@ storiesOf('Input')
 					defaultValue={inputData.tallText[0]}
 				/>
 				<Input
+					autoFocus={boolean('autoFocus')}
 					onChange={action('onChange')}
 					disabled={boolean('disabled')}
 					iconAfter={select('iconAfter', iconNames)}
@@ -58,6 +61,7 @@ storiesOf('Input')
 					defaultValue={inputData.tallText[1]}
 				/>
 				<Input
+					autoFocus={boolean('autoFocus')}
 					onChange={action('onChange')}
 					disabled={boolean('disabled')}
 					iconAfter={select('iconAfter', iconNames)}
@@ -74,6 +78,7 @@ storiesOf('Input')
 		'with extra spacing',
 		() => (
 			<Input
+				autoFocus={boolean('autoFocus')}
 				onChange={action('onChange')}
 				disabled={boolean('disabled')}
 				iconAfter={select('iconAfter', iconNames)}
@@ -91,6 +96,7 @@ storiesOf('Input')
 			<div>
 				<div>
 					<Input
+						autoFocus={boolean('autoFocus')}
 						onChange={action('onChange')}
 						disabled={boolean('disabled')}
 						iconAfter={select('iconAfter', iconNames)}
@@ -101,6 +107,7 @@ storiesOf('Input')
 						defaultValue={inputData.initialValue + ' one'}
 					/>
 					<Input
+						autoFocus={boolean('autoFocus')}
 						onChange={action('onChange')}
 						disabled={boolean('disabled')}
 						iconAfter={select('iconAfter', iconNames)}
@@ -113,6 +120,7 @@ storiesOf('Input')
 				</div>
 				<div>
 					<Input
+						autoFocus={boolean('autoFocus')}
 						onChange={action('onChange')}
 						disabled={boolean('disabled')}
 						iconAfter={select('iconAfter', iconNames)}
@@ -123,6 +131,7 @@ storiesOf('Input')
 						defaultValue={inputData.initialValue + ' three'}
 					/>
 					<Input
+						autoFocus={boolean('autoFocus')}
 						onChange={action('onChange')}
 						disabled={boolean('disabled')}
 						iconAfter={select('iconAfter', iconNames)}
@@ -140,6 +149,7 @@ storiesOf('Input')
 		'with a range',
 		() => (
 			<Input
+				autoFocus={boolean('autoFocus')}
 				onChange={action('onChange')}
 				disabled={boolean('disabled')}
 				iconAfter={select('iconAfter', iconNames)}